Initial refactoring to start implementing the new spamdetection strategy
