+++ spamdetector/__init__.py
