+++ spamdetector/mock.py
