+++ spamdetector/utils/config.py
