+++ spamdetector/utils/misc.py
