+++ tests/test_template_validator.py
@@ -24,11 +24,11 @@ import bittensor as bt
 from neurons.validator import Neuron as Validator
 from neurons.miner import Neuron as Miner
 
-from template.protocol import Dummy
-from template.validator.forward import forward
-from template.utils.uids import get_random_uids
-from template.validator.reward import get_rewards
-from template.base.validator import BaseValidatorNeuron
+from spamdetector.protocol import Dummy
+from spamdetector.validator.forward import forward
+from spamdetector.utils.uids import get_random_uids
+from spamdetector.validator.reward import get_rewards
+from spamdetector.base.validator import BaseValidatorNeuron
 
 
 class TemplateValidatorNeuronTestCase(unittest.TestCase):
